double'a çevrildi

+++ G03_CENG431_Final/src/factory/BankWalletFactory.java
@@ -71,7 +71,7 @@ public class BankWalletFactory extends WalletFactory {
 		
 		Object resultObject = result.getObject();
 		if(resultObject!=null){
-			entityResult = new WalletEntity((ICurrency) resultObject,Float.valueOf(quantity));
+			entityResult = new WalletEntity((ICurrency) resultObject,Double.valueOf(quantity));
 		}
 		return entityResult;
 	}
