+++ G03_CENG431_Final/src/model/WalletHelper.java
@@ -3,8 +3,8 @@ package model;
 import storage.IContainer;
 
 public class WalletHelper{
-	protected static float getTotal(IContainer<WalletEntity> entities, String banknoteName){
-		float result = 0;
+	protected static Double getTotal(IContainer<WalletEntity> entities, String banknoteName){
+		Double result = (double)0;
 		for(WalletEntity entity: entities){
 			result+=entity.getTotal(banknoteName);
 		}
