+++ G03_CENG431_Final/src/view/LabelInfo.java
@@ -6,10 +6,10 @@ public class LabelInfo {
 	public Color color;
 	public Double percent;
 	public String coinName;
-	public Float value;
+	public Double value;
 	public String banknote;
 
-	public LabelInfo(Color color, Double percent, String name, Float value, String banknote) {
+	public LabelInfo(Color color, Double percent, String name, Double value, String banknote) {
 		this.color = color;
 		this.percent = percent;
 		this.coinName = name;
